rebase update

+++ src/main/java/com/portale/mapper/NotificationMapper.java
@@ -13,20 +13,20 @@ public interface NotificationMapper {
 	Long GetNotificationsCount(@Param("user_id") int user_id);
 
 	void CreateNotification(@Param("notification") NotificationObject notification, @Param("date") Long date,
-			@Param("title") String title, @Param("message") String message, @Param("importancy") Long importancyLevel,
-			@Param("from_user") Long from_user, @Param("type") int type);
+			@Param("title") String title, @Param("message") String message, @Param("importancy") int importancyLevel,
+			@Param("from_user") int from_user, @Param("type") int type);
 
-	void AppendNotificationToUser(@Param("ref_to") Long ref_to, @Param("notification_ref") Long notification_ref);
+	void AppendNotificationToUser(@Param("ref_to") int ref_to, @Param("notification_ref") int notification_ref);
 
-	void SetNotificationAsSeen(@Param("user_id") Long user_id, @Param("notificationId") Long notificationId);
+	void SetNotificationAsSeen(@Param("user_id") int user_id, @Param("notificationId") int notificationId);
 
-	Integer isNotificationForUser(@Param("user_id") Long user_id, @Param("notificationId") Long notificationId);
+	Integer isNotificationForUser(@Param("user_id") int user_id, @Param("notificationId") int notificationId);
 
-	NotificationObject GetNotificationById(@Param("user_id") Long user_id,
-			@Param("notificationId") Long notificationId);
+	NotificationObject GetNotificationById(@Param("user_id") int user_id,
+			@Param("notificationId") int notificationId);
 
 	List<NotificationObject> GetAllNotifications(@Param("user_id") int user_id, @Param("limit") int limit,
 			@Param("offset") int offset);
 
-	Boolean AnyUnseenNotifications(@Param("to") Long userRequestLogin);
+	Boolean AnyUnseenNotifications(@Param("to") int userRequestLogin);
 }
